Set number of outcome columns based on screen width and remove OutcomeView
